pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yex-trader"
version = "0.1.0"
description = "Autonomous Hyperliquid trading CLI — deterministic and LLM-powered strategies for HL perps and YEX markets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nunchi Trade", email = "dev@nunchi.trade" },
]
keywords = ["hyperliquid", "trading", "market-making", "defi", "yex", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "hyperliquid-python-sdk>=0.4.0",
    "eth-account>=0.10.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40.0"]
mcp = ["mcp>=1.0.0"]
dev = ["pytest>=7.0", "ruff>=0.4.0", "mypy>=1.8.0"]

[project.scripts]
hl = "cli.main:main"

[project.urls]
Homepage = "https://github.com/Nunchi-trade/agent-cli"
Repository = "https://github.com/Nunchi-trade/agent-cli"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.setuptools.packages.find]
include = ["cli*", "strategies*", "sdk*", "common*", "parent*", "modules*", "skills*"]
